checkinfo should look for "Tags:" and not just "Tags"
A field named exactly "Tags:" must exist in the .info file. If not, the script that builds tags.db for the repository will not add the extension to  tags.db. If extension is not in tags.db, user will be unable to find the extension when searching for it using tce or Apps.